+++ src/index.ts
@@ -1,3 +1,4 @@
+import './theme.css'
 export const VERSION = '0.1.0'
 
 // Types
@@ -19,3 +20,11 @@ export { useWindowDrag } from './composables/useWindowDrag'
 // Components
 export { default as WindowFrame } from './components/WindowFrame.vue'
 export { default as WindowHost } from './components/WindowHost.vue'
+
+export * from './core'
+export { default as ExternalContent } from './components/ExternalContent.vue'
+export { default as MenuBar } from './components/MenuBar.vue'
+export { default as CommandToolbar } from './components/CommandToolbar.vue'
+export { default as CommandMenu } from './components/CommandMenu.vue'
+
+export { default as ContextMenu } from './components/ContextMenu.vue'